requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
scrapetube==2.5.1
youtube-transcript-api==0.6.2
yt-dlp==2024.8.6
openai-whisper==20231117
pydantic==2.9.2
python-dotenv==1.0.1
aiohttp==3.10.5
